dir restructure - forgot to migrate arxiv and gtr as well
