+++ nesta/core/scripts/docker_recipes/py37_amzn1_es7/Dockerfile
@@ -22,9 +22,9 @@ RUN yum clean all -y
 # Install base conda env
 RUN wget --quiet https://repo.anaconda.com/miniconda/Miniconda3-latest-Linux-x86_64.sh -O ~/miniconda.sh && /bin/bash ~/miniconda.sh -b -p /opt/conda && rm ~/miniconda.sh && ln -s /opt/conda/etc/profile.d/conda.sh /etc/profile.d/conda.sh && echo ". /opt/conda/etc/profile.d/conda.sh" >> ~/.bashrc && echo "conda activate base" >> ~/.bashrc && /bin/bash ~/.bashrc
 ENV PATH /opt/conda/bin:$PATH
-RUN conda create -y -n py36 \
+RUN conda create -y -n py37 \
     nomkl \
-    python=3.6 \
+    python=3.7 \
     cython \
     && conda clean -afy \
     && find /opt/conda/ -follow -type f -name '*.a' -delete \
